ソースのコピーライトを修正。 エイリアス関連の不具合を修正。 DMのコンパイルに関する不具合を修正。
git-svn-id: file:///home/moriyoshi/chaperone.e-cell.org/var/opt/subversion/ecell-ide/trunk@3585 fe1fbaf7-ad0a-4b3d-a902-8cb9c273ac6b

+++ ide/PathwayWindow/PathwayWindow/Animation/EdgeAnimationItem.cs
@@ -2,7 +2,7 @@
 //
 //        This file is part of E-Cell Environment Application package
 //
-//                Copyright (C) 1996-2006 Keio University
+//                Copyright (C) 1996-2010 Keio University
 //
 //::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 //
@@ -382,8 +382,8 @@ namespace Ecell.IDE.Plugins.PathwayWindow.Animation
         public override void ApplyChange()
         {
             _autoThreshold = autoThresholdCheckBox.Checked;
-            _thresholdHigh = float.Parse(thresholdHigh.Text);
-            _thresholdLow = float.Parse(thresholdLow.Text);
+            _thresholdHigh = double.Parse(thresholdHigh.Text);
+            _thresholdLow = double.Parse(thresholdLow.Text);
             _highEdgeBrush = edgeHighBrush.Brush;
             _lowEdgeBrush = edgeLowBrush.Brush;
             _ngEdgeBrush = edgeNGBrush.Brush;
